+++ src/components/app/App.tsx
@@ -4,13 +4,17 @@ import { Provider } from 'react-redux';
 
 import { BackToTop, Header } from '..';
 
+import { ThemeProvider } from '@mui/material/styles';
+import { CssBaseline } from '@mui/material';
+
+import theme from '../themeProvider/theme';
+
 import store from '../../reduxUsers/store';
-import Routing from './Routing';
+
+import Routing from '../routing/Routing';
 import ModalWindow from '../modal';
 import { Footer } from '../footer/Footer';
-import theme from '../themeProvider/theme';
-import { ThemeProvider } from '@mui/material/styles';
-import { Container, CssBaseline } from '@mui/material';
+import StyledMuiContainerAppMain from './StyledMuiContainer';
 
 function App() {
   return (
@@ -20,23 +24,10 @@ function App() {
         <HashRouter>
           <ModalWindow />
           <Header />
-          <Container
-            sx={{
-              padding: { xs: 0, md: 0 },
-              minHeight: {
-                xs: 'calc(100vh - 152px)',
-                sm: 'calc(100vh - 106px)',
-                md: 'calc(100vh - 110px)',
-              },
-              overflow: 'auto',
-            }}
-          >
+          <StyledMuiContainerAppMain>
             <Routing />
             <BackToTop />
-          </Container>
-          {/* <MainContainer> */}
-          {/* <Routing />
-          {/* </MainContainer> */}
+          </StyledMuiContainerAppMain>
           <Footer />
         </HashRouter>
       </Provider>
